requirements.txt
livekit-agents[openai,silero,turn-detector,mcp]>=1.0
livekit-api>=1.0
mcp[cli]>=1.8
pydantic>=2.0
python-dotenv>=1.0
twilio>=9.0
httpx>=0.27
opentelemetry-sdk>=1.20
opentelemetry-exporter-otlp-proto-http>=1.20
pytest>=8.0
pytest-asyncio>=0.24
